refactor: rename api package
