+++ internal/http/api/clients_test.go
@@ -1,4 +1,4 @@
-package controller_test
+package api_test
 
 import (
 	"bytes"
@@ -12,7 +12,7 @@ import (
 	"github.com/go-chi/chi/v5"
 	"github.com/go-faker/faker/v4"
 	"github.com/raphael-foliveira/chi-gorm/internal/exceptions"
-	"github.com/raphael-foliveira/chi-gorm/internal/http/controller"
+	"github.com/raphael-foliveira/chi-gorm/internal/http/api"
 	"github.com/raphael-foliveira/chi-gorm/internal/http/schemas"
 	"github.com/stretchr/testify/assert"
 )
@@ -21,7 +21,7 @@ func TestClient_List(t *testing.T) {
 	t.Run("should list all clients", testCase(func(t *testing.T, deps *testDependencies) {
 		recorder := httptest.NewRecorder()
 		request := httptest.NewRequest("GET", "/", nil)
-		ctx := controller.NewContext(recorder, request)
+		ctx := api.NewContext(recorder, request)
 		err := deps.clientsController.List(ctx)
 		assert.NoError(t, err)
 		assert.Equal(t, http.StatusOK, recorder.Code)
@@ -35,7 +35,7 @@ func TestClient_Get(t *testing.T) {
 		tx := chi.NewRouteContext()
 		tx.URLParams.Add("id", fmt.Sprintf("%d", deps.clientsStubs[0].ID))
 		request = request.WithContext(context.WithValue(request.Context(), chi.RouteCtxKey, tx))
-		ctx := controller.NewContext(recorder, request)
+		ctx := api.NewContext(recorder, request)
 		err := deps.clientsController.Get(ctx)
 		assert.NoError(t, err)
 		assert.Equal(t, http.StatusOK, recorder.Code)
@@ -49,7 +49,7 @@ func TestClient_Create(t *testing.T) {
 		faker.FakeData(&newClient)
 		reqBody, _ := json.Marshal(newClient)
 		request := httptest.NewRequest("POST", "/", bytes.NewReader(reqBody))
-		ctx := controller.NewContext(recorder, request)
+		ctx := api.NewContext(recorder, request)
 		err := deps.clientsController.Create(ctx)
 		assert.NoError(t, err)
 		assert.Equal(t, http.StatusCreated, recorder.Code)
@@ -59,7 +59,7 @@ func TestClient_Create(t *testing.T) {
 		invalidReqBody := `{"foo: 95}`
 		recorder := httptest.NewRecorder()
 		request := httptest.NewRequest("POST", "/", bytes.NewReader([]byte(invalidReqBody)))
-		ctx := controller.NewContext(recorder, request)
+		ctx := api.NewContext(recorder, request)
 		err := deps.clientsController.Create(ctx)
 		apiErr, ok := err.(*exceptions.ApiError)
 		assert.True(t, ok, "err should be an ApiError")
@@ -77,7 +77,7 @@ func TestClient_Update(t *testing.T) {
 		tx := chi.NewRouteContext()
 		tx.URLParams.Add("id", "1")
 		request = request.WithContext(context.WithValue(request.Context(), chi.RouteCtxKey, tx))
-		ctx := controller.NewContext(recorder, request)
+		ctx := api.NewContext(recorder, request)
 		err := deps.clientsController.Update(ctx)
 		assert.NoError(t, err)
 		assert.Equal(t, http.StatusOK, recorder.Code)
@@ -90,7 +90,7 @@ func TestClient_Update(t *testing.T) {
 		tx := chi.NewRouteContext()
 		tx.URLParams.Add("id", "1")
 		request = request.WithContext(context.WithValue(request.Context(), chi.RouteCtxKey, tx))
-		ctx := controller.NewContext(recorder, request)
+		ctx := api.NewContext(recorder, request)
 		err := deps.clientsController.Update(ctx)
 		apiErr, ok := err.(*exceptions.ApiError)
 		assert.True(t, ok, "err should be an ApiError")
@@ -105,7 +105,7 @@ func TestClient_Delete(t *testing.T) {
 		tx := chi.NewRouteContext()
 		tx.URLParams.Add("id", "1")
 		request = request.WithContext(context.WithValue(request.Context(), chi.RouteCtxKey, tx))
-		ctx := controller.NewContext(recorder, request)
+		ctx := api.NewContext(recorder, request)
 		err := deps.clientsController.Delete(ctx)
 		assert.NoError(t, err)
 		assert.Equal(t, http.StatusNoContent, recorder.Code)
