+++ internal/http/api/controller.go
@@ -1,4 +1,4 @@
-package controller
+package api
 
 import (
 	"encoding/json"
